Moving ViewModels out of Data namespace
