+++ Plugins/BTCPayServer.RockstarDev.Plugins.Payroll/Views/PayrollUser/List.cshtml
@@ -1,15 +1,9 @@
 ﻿@using BTCPayServer.Abstractions.Extensions
 @using BTCPayServer.Abstractions.Contracts
 @using BTCPayServer.Abstractions.Models
-@using BTCPayServer.Client
-@using BTCPayServer.RockstarDev.Plugins.Payroll.Controllers
-@using BTCPayServer.RockstarDev.Plugins.Payroll.Data.Models
 @using BTCPayServer.RockstarDev.Plugins.Payroll.Data
-@using BTCPayServer.Services
-@using static BTCPayServer.RockstarDev.Plugins.Payroll.Controllers.PayrollUserController
 @inject IScopeProvider ScopeProvider
-@inject DisplayFormatter DisplayFormatter
-@model PayrollUserListViewModel
+@model BTCPayServer.RockstarDev.Plugins.Payroll.ViewModels.PayrollUser.PayrollUserListViewModel
 @{
     ViewData.SetActivePage("Payroll", "Payroll Users", "Payroll");
